feat(imageData): Implement useContext imageId, imageTags and imageDescription are now passed through ImageDataContext, which significantly cleans up all components in the ReactivePictureFrame family!
